Rename MyProfile.js to Index.js
